+++ alarm/src/main/java/com/mdh/alarm/message/Message.java
@@ -1,0 +1,8 @@
+package com.mdh.alarm.message;
+
+public record Message(
+        String to,
+        String from,
+        String message
+) {
+}
