+++ alarm/src/main/java/com/mdh/alarm/message/MessageSender.java
@@ -1,0 +1,7 @@
+package com.mdh.alarm.message;
+
+import com.mysql.cj.Messages;
+
+public interface MessageSender {
+    void send(Messages message);
+}
