+++ alarm/src/main/java/com/mdh/alarm/slack/SlackAlertRequest.java
@@ -1,4 +1,4 @@
-package com.mdh.alarm.common.api;
+package com.mdh.alarm.slack;
 
 public record SlackAlertRequest(
         String text
